==> federated_learning/2_make_syn_data_by_total_data.py <==
import numpy as np
import pandas as pd
import syft as sy
from ctgan import CTGAN
import torch
import copy
from collections import OrderedDict
import warnings

warnings.filterwarnings("ignore", category=FutureWarning)


def initialize_device():
    # FIXME torch 버전 등 이슈로 cpu만 인식 중
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    print(f"Using device: {device}")
    return device


def load_data(file_path, num_samples=100000):
    data = pd.read_csv(file_path)[:num_samples]
    data = data.apply(pd.to_numeric, errors='coerce').fillna(0).astype(np.int64)
    return data

def train_ctgan(data, embedding_dim=128, generator_dim=(256, 256), discriminator_dim=(256, 256), pac=10):
    print("Data content (first 5 rows):")
    print(data[:5])

    if hasattr(data, 'child'):
        data = data.child

    if isinstance(data, pd.DataFrame):
        data_list = data.values
    elif isinstance(data, np.ndarray):
        # numpy array인 경우 DataFrame으로 변환
        data_list = pd.DataFrame(data.tolist(), columns=['BASE_YM', 'HNDE_BANK_RPTV_CODE', 'OPENBANK_RPTV_CODE', 'FND_TPCD', 'TRAN_AMT'])
    else:
        raise ValueError(f"Unexpected data type: {type(data)}")

    if len(data_list) == 0 or len(data_list[0]) != 5:
        raise ValueError(f"Data does not have the expected shape: {data_list[:5]}")

    columns = ['BASE_YM', 'HNDE_BANK_RPTV_CODE', 'OPENBANK_RPTV_CODE', 'FND_TPCD', 'TRAN_AMT']
    data_df = pd.DataFrame(data_list, columns=columns)

    model = CTGAN(
        embedding_dim=embedding_dim,
        generator_dim=generator_dim,
        discriminator_dim=discriminator_dim,
        pac=pac,
        epochs=10
    )
    print(model)

    # FIXME discrete_columns 설정 (OPENBANK_RPTV_CODE,FND_TPCD 추가)
    model.fit(data_df, discrete_columns=['HNDE_BANK_RPTV_CODE'])
    return model


if __name__ == "__main__":
    device = initialize_device()

    # 타행이체거래내역
    file_path = '../synthetic_data/org_datasets/DATOP_HF_TRANS_ENC_CODE.csv'
    data = load_data(file_path)

    model = train_ctgan(data)

    if model:
        synthetic_data = model.sample(620)
        synthetic_data['TRAN_AMT'] = synthetic_data['TRAN_AMT'].abs()
        synthetic_data.to_csv('data/synthetic_data_type1.csv', index=False)
        print("Synthetic Data Generated by the Federated Model:")
        print(synthetic_data)
    else:
        print("No models were successfully trained.")
